Move are_migrations_running function to file that isn't full of errors.
